Back to DejaVu all over again!
subtitle: How I learned to love Xresources

The main theme of these changes is to move back to DejaVu fonts
(ditching commitmono because they still don't offer default TTF fonts!)

Part of this was spurred by attempts to run XQuartz on a MacBook Air
with a 13-inch Retina display at 227 dpi (2880 x 1800).  Having to try
to find the commitmono TTF files from a reliable source again
(i.e. instead of copying the ones I was using from a third party) seemed
impossible now, so I re-evaluated alternatives and settled again on
DejaVu, which continues to be very well maintained and growing.

As per the subtitle, some default command-line options, i.e. mostly
XTERM_OPTS, that are moved out of environment variables and into
~/.Xdefaults, thus simplifying ~/.ctwmrc considerably (at least in terms
of the M4 macros it defines).  What remains are geometry options for
specific classes or instances of apps such as Xterm and Xload, and these
are mostly set up in ~/.xinitrc and used therein and in ~/.ctwmrc, with
the messy parts being CTWM parameters (which cannot be from env vars).

BTW, changing '-n OccupyAll' to '-T OccupyAll' for xterm fixed the setup
of the right-hand xterms that should appear in all CTWM workspaces!!!

There are also some re-factoring of some of the M4 macros used with
xrdb(1), i.e. for .Xdefaults and .fonts.alias.m4.  (Some of these macros
are still duplicated in ~/.ctwmrc because of its different comment
syntax, sigh.)   Note the previous commit adding .X11-m4macros.m4!

There are some minor fixes to font settings, other resource settings,
etc., including using a macro name for the default "DecentFont"!

Get rid of use of "-display" wherever onx11server is used.

Also add Xchrono as a default app, if it is available.

Note there was also a change required to my ShortCuts shortcut for
returning to XQuartz due to the now dynamic nature of the screen size
and the fact that ShortCuts are shared across machines using the same
AppleID.